Change LanesInUse type to size_t (#628)
Change LanesInUse to type size_t from int64_t. This change is
made as per upstream changes at [1]
[1]: https://gerrit.openbmc.org/c/openbmc/phosphor-dbus-interfaces/+/53650

+++ host-bmc/dbus/custom_dbus.cpp
@@ -167,7 +167,7 @@ void CustomDBus::implementPCIeDeviceInterface(const std::string& path)
     }
 }
 
-void CustomDBus::setPCIeDeviceProps(const std::string& path, int64_t lanesInuse,
+void CustomDBus::setPCIeDeviceProps(const std::string& path, size_t lanesInuse,
                                     const std::string& value)
 {
     Generations generationsInuse =
